Reorganize internals: prev event_handlers
