+++ lib/action/core/flow/messages.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require "action/core/event_handlers"
+require "action/core/flow/handlers"
 
 module Action
   module Core
@@ -8,7 +8,7 @@ module Action
       module Messages
         def self.included(base)
           base.class_eval do
-            class_attribute :_messages_registry, default: Action::EventHandlers::Registry.empty
+            class_attribute :_messages_registry, default: Action::Core::Flow::Handlers::Registry.empty
 
             extend ClassMethods
             include InstanceMethods
@@ -26,7 +26,7 @@ module Action
             _messages_registry.for(event_type).each do |handler|
               next if handler.respond_to?(:static?) && handler.static?
 
-              msg = handler.execute_if_matches(action:, exception:)
+              msg = handler.apply(action:, exception:)
               return msg if msg.present?
             end
             nil
@@ -36,15 +36,10 @@ module Action
             _messages_registry.for(event_type).each do |handler|
               next unless handler.respond_to?(:static?) && handler.static?
 
-              msg = handler.execute_if_matches(action:, exception:)
+              msg = handler.apply(action:, exception:)
               return msg if msg.present?
             end
             nil
-          end
-
-          # Internal introspection helper
-          def _messages_for(event_type)
-            Array(_messages_registry.for(event_type))
           end
 
           def success(message = nil, **kwargs, &block)
@@ -74,7 +69,7 @@ module Action
 
           # Private helpers
           def _add_message(kind, msg, matcher:, static:)
-            entry = Action::EventHandlers::MessageHandler.new(matcher:, message: msg, static:)
+            entry = Action::Core::Flow::Handlers::MessageHandler.new(matcher:, message: msg, static:)
             self._messages_registry = _messages_registry.register(event_type: kind, entry:, prepend: true)
           end
           private :_add_message
